Adjusted task 4

+++ Labs/Lab_3/Date.cs
@@ -51,7 +51,7 @@
             Month += monthAdded;
             Day += howMany;
 
-            if ((int)Month > 12)
+            if (Month > 12)
             {
                 Year++;
                 Month -= 12;
